+++ packages/react/src/components/index.ts
@@ -127,6 +127,8 @@ export * from './DialogContentText';
 export {default as DialogTitle} from './DialogTitle';
 export * from './DialogTitle';
 
+export * from './dnd';
+
 export {default as Divider} from './Divider';
 export * from './Divider';
 
@@ -249,6 +251,27 @@ export * from './Switch';
 
 export {default as Tab} from './Tab';
 export * from './Tab';
+
+export {default as Table} from './Table';
+export * from './Table';
+
+export {default as TableBody} from './TableBody';
+export * from './TableBody';
+
+export {default as TableCell} from './TableCell';
+export * from './TableCell';
+
+export {default as TableContainer} from './TableContainer';
+export * from './TableContainer';
+
+export {default as TableFooter} from './TableFooter';
+export * from './TableFooter';
+
+export {default as TableHead} from './TableHead';
+export * from './TableHead';
+
+export {default as TableRow} from './TableRow';
+export * from './TableRow';
 
 export {default as TabPanel} from './TabPanel';
 export * from './TabPanel';